[refact] seperate infinite FeedList
