+++ src/app/(MainLayout)/components/FeedBottom/FeedBottom.tsx
@@ -4,13 +4,14 @@ import styles from './FeedBottom.module.scss';
 
 type FeedBottomProps = {
   feed: Feed.FeedData;
+  setDetailFeedId: (feedId: number) => void;
 };
 
-const FeedBottom = ({ feed }: FeedBottomProps) => {
+const FeedBottom = ({ feed, setDetailFeedId }: FeedBottomProps) => {
   return (
     <div className={styles.feed_bottom}>
-      <FeedCountBox post={feed.post} />
-      <FeedContentBox feed={feed} />
+      <FeedCountBox setDetailFeedId={setDetailFeedId} post={feed.post} />
+      <FeedContentBox setDetailFeedId={setDetailFeedId} feed={feed} />
     </div>
   );
 };
