+++ src/app/(MainLayout)/components/FeedCountBox/FeedCountBox.tsx
@@ -8,12 +8,14 @@ import styles from './FeedCountBox.module.scss';
 
 type FeedCountBoxProps = {
   post: Feed.BasicPost | Feed.DebatePost | Feed.PollPost;
+  setDetailFeedId: (feedId: number) => void;
 };
 
-const FeedCountBox = ({ post }: FeedCountBoxProps) => {
+const FeedCountBox = ({ post, setDetailFeedId }: FeedCountBoxProps) => {
   const router = useRouter();
 
   const onClickMessageIcon = () => {
+    setDetailFeedId(post.postId);
     if (post.postType === 'BASIC') {
       router.push('?cf=1');
     }
